+++ tools/quake3/q3map2/visflow.c
@@ -1,4 +1,5 @@
-/*
+/* -------------------------------------------------------------------------------
+
 Copyright (C) 1999-2007 id Software, Inc. and contributors.
 For a list of contributors, see the accompanying CONTRIBUTORS file.
 
@@ -1400,7 +1401,7 @@ void CreatePassages(int portalnum)
 			/* ydnar: prefer correctness to stack overflow  */
 			//% memcpy( &in, p->winding, (int)((fixedWinding_t *)0)->points[p->winding->numpoints] );
 			if( p->winding->numpoints <= MAX_POINTS_ON_FIXED_WINDING )
-				memcpy( &in, p->winding, (int) &(((fixedWinding_t*) 0)->points[ p->winding->numpoints ]) );
+				memcpy( &in, p->winding, (size_t) &(((fixedWinding_t*) 0)->points[ p->winding->numpoints ]) );
 			else
 				memcpy( &in, p->winding, sizeof( fixedWinding_t ) );
